override Preface::getDMSchemes to allow missing property by default
Files produced by EVS were missing the required DMSchemes property.
This change logs a warning and returns an empty array in that case.

The allow_missing method parameter defaults to true, but can be set
to false to have the previous behaviour of throwing an exception
when the property is missing.